+++ src/azure-cli-core/azure/cli/core/api.py
@@ -8,3 +8,6 @@
 # pylint: disable=unused-import
 
 from ._environment import get_config_dir
+from ._profile import load_subscriptions
+
+from ._completers import get_subscription_id_list