+++ src/azure-cli-core/azure/cli/core/commands/arm.py
@@ -21,7 +21,7 @@ from azure.cli.core.commands import LongRunningOperation, _is_poller
 from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.commands.validators import IterateValue
 from azure.cli.core.util import shell_safe_json_parse, augment_no_wait_handler_args, get_command_type_kwarg
-from azure.cli.core.profiles import ResourceType
+from azure.cli.core.profiles import ResourceType, get_sdk
 
 logger = get_logger(__name__)
 EXCLUDED_NON_CLIENT_PARAMS = list(set(EXCLUDED_PARAMS) - set(['self', 'client']))
@@ -298,7 +298,7 @@ def register_global_subscription_parameter(cli_ctx):
     import knack.events as events
 
     def add_subscription_parameter(_, **kwargs):
-        from azure.cli.command_modules.profile._completers import get_subscription_id_list
+        from azure.cli.core._completers import get_subscription_id_list
 
         commands_loader = kwargs['commands_loader']
         cmd_tbl = commands_loader.command_table
@@ -698,13 +698,17 @@ def _cli_show_command(context, name, getter_op, custom_command=False, **kwargs):
         try:
             return getter(**args)
         except Exception as ex:  # pylint: disable=broad-except
-            if getattr(getattr(ex, 'response', ex), 'status_code', None) == 404:
-                logger.error(getattr(ex, 'message', ex))
-                import sys
-                sys.exit(3)
-            raise
+            show_exception_handler(ex)
     context._cli_command(name, handler=handler, argument_loader=generic_show_arguments_loader,  # pylint: disable=protected-access
                          description_loader=description_loader, **kwargs)
+
+
+def show_exception_handler(ex):
+    if getattr(getattr(ex, 'response', ex), 'status_code', None) == 404:
+        logger.error(getattr(ex, 'message', ex))
+        import sys
+        sys.exit(3)
+    raise ex
 
 
 def verify_property(instance, condition):
@@ -984,8 +988,6 @@ def _find_property(instance, path):
 
 def assign_identity(cli_ctx, getter, setter, identity_role=None, identity_scope=None):
     import time
-    from azure.mgmt.authorization import AuthorizationManagementClient
-    from azure.mgmt.authorization.models import RoleAssignmentCreateParameters
     from msrestazure.azure_exceptions import CloudError
 
     # get
@@ -997,7 +999,10 @@ def assign_identity(cli_ctx, getter, setter, identity_role=None, identity_scope=
         principal_id = resource.identity.principal_id
 
         identity_role_id = resolve_role_id(cli_ctx, identity_role, identity_scope)
-        assignments_client = get_mgmt_service_client(cli_ctx, AuthorizationManagementClient).role_assignments
+        assignments_client = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_AUTHORIZATION).role_assignments
+        RoleAssignmentCreateParameters = get_sdk(cli_ctx, ResourceType.MGMT_AUTHORIZATION,
+                                                 'RoleAssignmentCreateParameters', mod='models',
+                                                 operation_group='role_assignments')
         parameters = RoleAssignmentCreateParameters(role_definition_id=identity_role_id, principal_id=principal_id)
 
         logger.info("Creating an assignment with a role '%s' on the scope of '%s'", identity_role_id, identity_scope)
@@ -1024,8 +1029,7 @@ def assign_identity(cli_ctx, getter, setter, identity_role=None, identity_scope=
 
 def resolve_role_id(cli_ctx, role, scope):
     import uuid
-    from azure.mgmt.authorization import AuthorizationManagementClient
-    client = get_mgmt_service_client(cli_ctx, AuthorizationManagementClient).role_definitions
+    client = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_AUTHORIZATION).role_definitions
 
     role_id = None
     if re.match(r'/subscriptions/[^/]+/providers/Microsoft.Authorization/roleDefinitions/',