+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,9 +3,13 @@
 Release History
 ===============
 
-2.2.1
+2.2.2
 +++++
 * `dns`: Added dns support to 2017-03-09-profile for Azure Stack 
+
+2.2.1
+++++++
+* Minor fixes
 
 2.2.0
 +++++