+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+2.1.2
+++++++
+* support for stack profile 2017-03-09-profile
+
+2.1.1
++++++
+* `ad app update`: Fixes issue where generic update parameters would not work correctly.
+
 2.1.0
 +++++
 * BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.